Move the registration text preview to its own component
