+++ wp-modules/editor/js/src/components/SidebarPanels/CategoriesPanel.tsx
@@ -2,7 +2,7 @@ import { __ } from '@wordpress/i18n';
 import { PluginDocumentSettingPanel } from '@wordpress/edit-post';
 import { Spinner } from '@wordpress/components';
 import Creatable from 'react-select/creatable';
-import CategoryRegistration from './Tertiary/CategoryRegistration';
+import CategoryRegistration from './CategoryRegistration';
 import convertToSlug from '../../utils/convertToSlug';
 
 import type { BaseSidebarProps, AdditionalSidebarProps } from './types';
